+++ test/test-app.js
@@ -6,14 +6,14 @@ var assert = require('yeoman-generator').assert;
 var helpers = require('yeoman-generator').test;
 var os = require('os');
 
-describe('nikita:app', function () {
+describe('nikita:app:custom-libsass', function () {
   before(function (done) {
     helpers.run(path.join(__dirname, '../app'))
       .inDir(path.join(os.tmpdir(), './temp-test'))
-      .withOptions({ 'skip-install': true })
+      .withOptions({ 'skipInstall': true })
       .withPrompt({
 		template: "custom",
-        sassCompiler: "compass",
+        sassCompiler: "libsass",
         private: true,
         name: "testrun" + (new Date()).getTime(),
         nikitaCssMixins: [],
@@ -31,4 +31,77 @@ describe('nikita:app', function () {
       'source/sass/styles.scss'
     ]);
   });
+});
+
+describe('nikita:app:custom-compass', function () {
+	before(function (done) {
+		helpers.run(path.join(__dirname, '../app'))
+			.inDir(path.join(os.tmpdir(), './temp-test'))
+			.withOptions({ 'skipInstall': true })
+			.withPrompt({
+				template: "custom",
+				sassCompiler: "compass",
+				private: true,
+				name: "testrun" + (new Date()).getTime(),
+				nikitaCssMixins: [],
+				useBuildFolders: true,
+				nikitaCssExtends: [],
+				features: []
+			})
+			.on('end', done);
+	});
+
+	it('creates files', function () {
+		assert.file([
+			'bower.json',
+			'package.json',
+			'source/sass/styles.scss'
+		]);
+	});
+});
+
+describe('nikita:app:slim', function () {
+	before(function (done) {
+		helpers.run(path.join(__dirname, '../app'))
+			.inDir(path.join(os.tmpdir(), './temp-test'))
+			.withOptions({ 'skipInstall': true })
+			.withPrompt({
+				template: "slim",
+				private: true,
+				useBuildFolders: true,
+				name: "testrun" + (new Date()).getTime()
+			})
+			.on('end', done);
+	});
+
+	it('creates files', function () {
+		assert.file([
+			'bower.json',
+			'package.json',
+			'source/sass/styles.scss'
+		]);
+	});
+});
+
+describe('nikita:app:default', function () {
+	before(function (done) {
+		helpers.run(path.join(__dirname, '../app'))
+			.inDir(path.join(os.tmpdir(), './temp-test'))
+			.withOptions({ 'skipInstall': true })
+			.withPrompt({
+				template: "default",
+				private: true,
+				useBuildFolders: true,
+				name: "testrun" + (new Date()).getTime()
+			})
+			.on('end', done);
+	});
+
+	it('creates files', function () {
+		assert.file([
+			'bower.json',
+			'package.json',
+			'source/sass/styles.scss'
+		]);
+	});
 });